requirements.txt
# ── Web framework ──────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
aiofiles==24.1.0

# ── Database ────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# ── Auth ────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# ── Redis / Queue ───────────────────────────────────────
redis==5.2.0
arq==0.26.1

# ── Scheduler ───────────────────────────────────────────
apscheduler==3.10.4

# ── HTTP client ─────────────────────────────────────────
httpx==0.28.0
aiohttp==3.11.9

# ── SEFAZ / XML ─────────────────────────────────────────
lxml==5.3.0
xmltodict==0.14.2
zeep==4.3.1
signxml==3.2.2
cryptography==43.0.3
pyOpenSSL==24.3.0

# ── Validation / Serialization ──────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# ── Utils ───────────────────────────────────────────────
python-dotenv==1.0.1
tenacity==9.0.0
structlog==24.4.0
sentry-sdk[fastapi]==2.19.0

# ── Dev / Test ──────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0
factory-boy==3.3.1
